Split off into components and constants and import into App
