+++ src/constants/breadcrumbNames.js
@@ -1,0 +1,8 @@
+export const BREADCRUMB_NAMES = {
+  '/users': 'Users',
+  '/users/1': '1',
+  '/users/2': '2',
+  '/users/1/detail': 'Detail',
+  '/users/2/detail': 'Detail',
+  '/about-us': 'About Us',
+};
